change listener and dispatch job
