+++ app/Providers/EventServiceProvider.php
@@ -7,7 +7,7 @@ use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Event;
 use App\Events\VideoUploaded;
-use App\Listeners\GenerateVideoThumbnail;
+use App\Listeners\VideoUploadedListener;
 
 class EventServiceProvider extends ServiceProvider
 {
@@ -21,7 +21,7 @@ class EventServiceProvider extends ServiceProvider
             SendEmailVerificationNotification::class,
         ],
         VideoUploaded::class => [
-            GenerateVideoThumbnail::class,
+            VideoUploadedListener::class,
         ]
     ];
